+++ packages/sdk/src/api-functions/generateDepositBtcAddress/generateDepositBtcAddress.ts
@@ -71,8 +71,8 @@ export interface IGenerateDepositBtcAddressParams extends IEnvParam {
   signatureData?: string;
 }
 
-const EXTRA_PARAMS_TOKENS = [Token.LBTC, Token.BTCK];
-const SUPPORTED_TOKENS = [Token.LBTC, Token.BTCK];
+const EXTRA_PARAMS_TOKENS = [Token.LBTC, Token.BTCK, Token.NativeLBTC];
+const SUPPORTED_TOKENS = [Token.LBTC, Token.BTCK, Token.NativeLBTC];
 
 /**
  * Generates a BTC deposit address.
@@ -150,7 +150,7 @@ export async function generateDepositBtcAddress({
         ] as BlockchainIdentifier[]
       ).includes(toChain)
     ) {
-      const tokenContractInfo = getTokenContractInfo(
+      const tokenContractInfo = await getTokenContractInfo(
         token,
         chainId as ChainId,
         env,