+++ packages/sdk/src/index.ts
@@ -30,7 +30,21 @@ export * from './utils/satoshi';
 
 // Metrics:
 export { getLBTCStats } from './metrics/get-lbtc-stats';
-export { getRewardsInfo } from './metrics/get-rewards-info';
+
+export {
+  getPositionsSummary,
+  type PositionsSummary,
+} from './metrics/get-positions-summary';
+export {
+  getApy,
+  type LbtcApy,
+  getEstimatedApy,
+  type LbtcEstimatedApy,
+} from './metrics/get-lbtc-apy';
+export {
+  getAdditionalRewards,
+  type RewardsDistribution,
+} from './metrics/get-additional-rewards';
 
 // Tokens:
 export * from './tokens/lbtc-addresses';
@@ -44,6 +58,10 @@ export type { Address, EIP1193Provider } from 'viem';
 export {
   bridge,
   bridgeCCIP,
+  bridgeOFT,
+  OFT_HI_GAS_LIMIT_CHAINS,
+  OFT_GAS_LIMIT,
+  OFT_HI_GAS_LIMIT,
   getBridgeInfo,
   type BridgeCCIPParameters,
   type BridgeParameters,